+++ instrumentation/active_job/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release History: opentelemetry-instrumentation-active_job
 
+### v0.7.8 / 2024-10-24
+
+* FIXED: ActiveJob Propagate baggage information properly when performing
+
+### v0.7.7 / 2024-08-21
+
+* FIXED: Propagate context between enqueue and perform
+
+### v0.7.6 / 2024-08-15
+
+* FIXED: Prefix ::ActiveSupport when installing the instrumentation
+
+### v0.7.5 / 2024-08-15
+
+* FIXED: Use Active Support Lazy Load Hooks to avoid prematurely initializing ActiveRecord::Base and ActiveJob::Base
+
 ### v0.7.4 / 2024-07-30
 
 * FIXED: Honour dynamic changes in configuration